// src/model/baidu-file.ts
export default interface BaiduFile {
  category: number;
  unlist: number;
  fs_id: number;
  dir_empty: number;
  oper_id: number;
  server_ctime: number;
  local_mtime: number;
  size: number;
  share: number;
  isdir: number;
  path: string;
  local_ctime: number;
  server_filename: string;
  empty: number;
  server_mtime: number;
}