// src/model/baidu-user.ts
import BaiduFile from './baidu-file';
export default interface BaiduUser {
  loginstate: number;
  username: string;
  third: number;
  flag: number;
  file_list?: BaiduFile[];
  uk: number;
  task_key: string;
  task_time: number;
  bdstoken: string;
  is_vip: number;
  bt_paths?: string;
  applystatus: number;
  sign1: string;
  sign2: string;
  sign3: string;
  sign4: string;
  sign5: string;
  timestamp: number;
  timeline_status: number;
  face_status: number;
  srv_ts: number;
  is_year_vip: number;
  show_vip_ad: number;
  is_evip: number;
  is_svip: number;
  is_auto_svip: number;
  activity_status: number;
  photo: string;
  curr_activity_code: number;
  activity_end_time: number;
  token: string;
  sharedir: number;
  pansuk: string;
  skinName: string;
  urlparam: string[];
  XDUSS: string;
}